fix(httpsig): repair CI (wolfSSL Ed25519, setopt, xmllint, REUSE, spacecheck)
- WolfSSL: compile Ed25519 signing only when import/sign APIs exist; else stub
- setopt: set/clear CURLAUTH_HTTPSIG in uint32_t space for -Wconversion
- test5011: %AMP in URL/protocol so xmllint accepts the testcase XML
- REUSE: annotate httpsig fixture keys; spacecheck: ASCII-only comments
  in httpsig_crypto.h

+++ lib/setopt.c
@@ -1115,9 +1115,9 @@ static CURLcode setopt_long_http(struct Curl_easy *data, CURLoption option,
   case CURLOPT_HTTPSIG:
     s->httpsig = arg;
     if(arg)
-      s->httpauth = CURLAUTH_HTTPSIG;
+      s->httpauth = (uint32_t)CURLAUTH_HTTPSIG;
     else
-      s->httpauth &= (unsigned long)~CURLAUTH_HTTPSIG;
+      s->httpauth &= ~(uint32_t)CURLAUTH_HTTPSIG;
     break;
 #endif
   default:
